reduced Duck Strategy exercise to fewer files
