+++ Design-Patterns/Strategy-Pattern/QuackBehaviour.cs
@@ -10,4 +10,28 @@ namespace Strategy_Pattern
     {
         void Quack();
     }
+
+    internal class Quacking : QuackBehaviour
+    {
+        public void Quack()
+        {
+            Console.WriteLine($"The duck quacked. 'Quack!'");
+        }
+    }
+
+    internal class MuteQuack : QuackBehaviour
+    {
+        public void Quack()
+        {
+            Console.WriteLine($"This duck can't quack.");
+        }
+    }
+
+    internal class Squeak : QuackBehaviour
+    {
+        public void Quack()
+        {
+            Console.WriteLine($"The duck squeaked. 'Squeak!'");
+        }
+    }
 }
