--- Design-Patterns/Strategy-Pattern/QuackBehaviours/MuteQuack.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Strategy_Pattern.QuackBehaviours
-{
-    internal class MuteQuack : QuackBehaviour
-    {
-        public void Quack()
-        {
-            Console.WriteLine($"This duck can't quack.");
-        }
-    }
-}
